Add public constants class for typedefs and definitions. Fix framework imports.
