+++ MSAL/src/MSAL_Internal.h
@@ -78,8 +78,8 @@
 typedef NSOrderedSet<NSString *> MSALScopes;
 
 #import "IdentityCore_Internal.h"
-#include "MSAL.h"
 #include "MSIDLogger+Internal.h"
 #include "MSALRequestParameters.h"
 #include "MSALError_Internal.h"
 #import  "MSIDRequestContext.h"
+#import "MSALConstants.h"
