+++ MSAL/src/public/MSALPublicClientApplication.h
@@ -26,7 +26,6 @@
 //------------------------------------------------------------------------------
 
 #import <Foundation/Foundation.h>
-#import <MSAL/MSAL.h>
 
 @class MSALResult;
 @class MSALAccount;
