+++ MSAL/test/app/ios/MSALTestAppUserViewController.m
@@ -30,6 +30,7 @@
 #import "MSALTestAppSettings.h"
 #import "MSALAccountId.h"
 #import "MSIDAuthorityFactory.h"
+#import "MSALAccount.h"
 
 @interface MSALTestAppUserViewController ()
